chore: move `cleanMetricName` to a shared helper file
